+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Fix] Timezone location lookup follows Debian/Ubuntu zoneinfo symlinks.
+
 ## [0.3.0] - 2026-08-16
 
 ### Added